requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
Deprecated==1.2.14
frozenlist==1.4.0
future==0.18.3
h11==0.14.0
idna==3.4
line-bot-sdk==3.5.0
lxml==4.9.3
multidict==6.0.4
numpy==1.26.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tabulate==0.9.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
webdriver-manager==4.0.1
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
